resolved descriptor error, optimised rust code
