+++ ios/Classes/bridge_generated.h
@@ -17,10 +17,6 @@ typedef int64_t DartPort;
 
 typedef bool (*DartPostCObjectFnType)(DartPort port_id, void *message);
 
-void wire_create_descriptor(int64_t port_, struct wire_uint_8_list *xprv);
-
-void wire_create_change_descriptor(int64_t port_, struct wire_uint_8_list *xprv);
-
 void wire_handle_rust(int64_t port_,
                       struct wire_uint_8_list *function,
                       struct wire_uint_8_list *arguments);
@@ -33,8 +29,6 @@ void store_dart_post_cobject(DartPostCObjectFnType ptr);
 
 static int64_t dummy_method_to_enforce_bundling(void) {
     int64_t dummy_var = 0;
-    dummy_var ^= ((int64_t) (void*) wire_create_descriptor);
-    dummy_var ^= ((int64_t) (void*) wire_create_change_descriptor);
     dummy_var ^= ((int64_t) (void*) wire_handle_rust);
     dummy_var ^= ((int64_t) (void*) new_uint_8_list_0);
     dummy_var ^= ((int64_t) (void*) free_WireSyncReturnStruct);
