+++ rust/log.txt
@@ -1,0 +1,2 @@
+2022-07-14-04:00 ///failed to parse arguments
+, {}: "err"
